requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiostream==0.4.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///opt/concourse/worker/volumes/live/8fbed55e-0573-4246-7cf9-162067141d5e/volume/cffi_1605538070636/work
chardet @ file:///opt/concourse/worker/volumes/live/a5a7c56b-23cb-471d-7404-f4130b8aed33/volume/chardet_1605303184529/work
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
conda==4.10.3
conda-package-handling @ file:///opt/concourse/worker/volumes/live/a7e34989-4c54-4cb6-4156-4e58ee270730/volume/conda-package-handling_1603018121300/work
cryptography @ file:///opt/concourse/worker/volumes/live/62f9527a-361c-47e5-5709-227a4a523add/volume/cryptography_1605544472602/work
cycler==0.11.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1626696982231/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1623610058511/work
docker==5.0.2
docopt==0.6.2
faiss==1.7.1
fastapi==0.68.1
filelock==3.0.12
Flask==2.0.1
fonttools==4.29.1
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1630424910447/work
grip==4.5.2
grpcio==1.40.0
h11==0.12.0
httptools==0.2.0
huggingface-hub==0.0.17
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1630267509190/work
itsdangerous==2.0.1
jina==2.1.1
Jinja2==3.0.1
joblib==1.1.0
kiwisolver==1.3.2
lz4==3.1.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.1
mkl-fft==1.3.0
mkl-random @ file:///opt/concourse/worker/volumes/live/f196a661-8e33-4aac-63bc-efb2ff50e035/volume/mkl_random_1626186080069/work
mkl-service==2.4.0
multidict==5.1.0
multiprocess @ file:///Users/runner/miniforge3/conda-bld/multiprocess_1623774555562/work
nltk==3.7
numpy==1.21.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1625323647219/work
pandas==1.4.1
path-and-address==2.0.1
pathspec==0.9.0
Pillow==8.3.2
protobuf==3.18.0
pyarrow==3.0.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydantic==1.8.2
Pygments==2.9.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
regex==2022.1.18
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
rich==10.10.0
ruamel-yaml==0.15.87
sacremoses==0.0.45
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
six @ file:///opt/concourse/worker/volumes/live/5b31cb27-1e37-4ca5-6e9f-86246eb206d2/volume/six_1605205320872/work
sklearn==0.0
starlette==0.14.2
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.3
# Editable Git install with no remote (tradingbot==0.1)
-e /Users/saileshnankani/Desktop/git-repos/rohan/trading_platform
transformers==4.10.2
typing-extensions==3.10.0.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.1
xxhash @ file:///Users/runner/miniforge3/conda-bld/python-xxhash_1618590156407/work
yarl==1.6.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1625284368454/work
